requirements.txt
anthropic
playwright
gspread
google-auth
google-auth-oauthlib
python-dotenv
requests
